Move DataTemplateSelector to separate file
