+++ 9.0/Apps/DeveloperBalance/DeveloperBalance/Pages/Controls/ChipDataTemplateSelector.cs
@@ -1,0 +1,12 @@
+namespace DeveloperBalance.Pages.Controls;
+
+public class ChipDataTemplateSelector : DataTemplateSelector
+{
+	public required DataTemplate SelectedTagTemplate { get; set; }
+	public required DataTemplate NormalTagTemplate { get; set; }
+
+	protected override DataTemplate OnSelectTemplate(object item, BindableObject container)
+	{
+		return (item as Tag)?.IsSelected ?? false ? SelectedTagTemplate : NormalTagTemplate;
+	}
+}
