+++ 9.0/Apps/DeveloperBalance/DeveloperBalance/Pages/ProjectDetailPage.xaml.cs
@@ -11,14 +11,3 @@ public partial class ProjectDetailPage : ContentPage
 		BindingContext = model;
 	}
 }
-
-public class ChipDataTemplateSelector : DataTemplateSelector
-{
-	public required DataTemplate SelectedTagTemplate { get; set; }
-	public required DataTemplate NormalTagTemplate { get; set; }
-
-	protected override DataTemplate OnSelectTemplate(object item, BindableObject container)
-	{
-		return (item as Tag)?.IsSelected ?? false ? SelectedTagTemplate : NormalTagTemplate;
-	}
-}
